store refactoring
