+++ src/store/actions/index.js
@@ -1,0 +1,2 @@
+import { createAction } from '@reduxjs/toolkit';
+export const reset = createAction('all/reset');
